1. Rename MetadataParser to MetadataContentParser and MetadataManager to MetadataParser 2. Add TODOs and Deprecate non-generic methods
