+++ server/src/main/java/org/opensearch/index/store/remote/metadata/RemoteSegmentMetadataContentParser.java
@@ -13,12 +13,12 @@ import java.util.Map;
 
 import org.apache.lucene.store.IndexInput;
 import org.apache.lucene.store.IndexOutput;
-import org.opensearch.common.metadata.MetadataParser;
+import org.opensearch.common.metadata.MetadataContentParser;
 
 /**
  * Parser for {@link RemoteSegmentMetadata}
  */
-public class RemoteSegmentMetadataParser implements MetadataParser<RemoteSegmentMetadata> {
+public class RemoteSegmentMetadataContentParser implements MetadataContentParser<RemoteSegmentMetadata> {
     /**
      * Reads metadata content from metadata file input stream and parsed into {@link RemoteSegmentMetadata}
      * @param indexInput metadata file input stream with {@link IndexInput#getFilePointer()} pointing to metadata content
@@ -41,9 +41,6 @@ public class RemoteSegmentMetadataParser implements MetadataParser<RemoteSegment
      * Writes metadata to file output stream
      * @param indexOutput metadata file input stream
      * @param content metadata content
-     *
-     * This method would be removed in future
-     * and only {@link #writeContent(IndexOutput, RemoteSegmentMetadata)} would be leveraged in future
      */
     @Override
     public void writeContent(IndexOutput indexOutput, Map<String, String> content) throws IOException {
